+++ CHANGES.md
@@ -4,6 +4,7 @@
 ## 0.1.1 (2023-MM-DD)
 
 - **BREAKING CHANGE**: Switching from `pivot`/`child` to `reference`/`secondary`
+- **BREAKING CHANGE**: Switching from `ruling` to `reference`
 - FIX: Fix Pair without secondary
 - CI: Test Pair without secondary and Mosaic and Series with only one product